Make class names for nodes consistent
